fix type issues with adding question
