+++ src/actions/addQuestion.ts
@@ -3,14 +3,13 @@
 import {
   addQuestionFormSchema,
   EnumAddQuestionResult,
+  EnumQuestionType,
 } from "@/schemas/questionSchema";
 import { EnumAccessType } from "@/schemas/dbTableAccessSchema";
 import { auth } from "@/utils/auth/config";
-import { addQuestionToDb } from "@/utils/classroom/addQuestionToDb";
 import { canUserAccessAssignment } from "@/utils/classroom/canUserAccessAssignment";
-import { putObject } from "@/utils/storage/s3/putObject";
-import { randomUUID } from "crypto";
 import { createSafeActionClient } from "next-safe-action";
+import { addCodeQuestion } from "@/utils/classroom/question/addCodeQuestion";
 
 export const addQuestion = createSafeActionClient()
   .schema(addQuestionFormSchema)
@@ -22,9 +21,8 @@ export const addQuestion = createSafeActionClient()
         return { type: EnumAddQuestionResult.NotAuthorized };
       }
 
-      const { question, name, assignmentId, starterCode } = parsedInput;
       const isAuthorized = await canUserAccessAssignment({
-        assignmentId,
+        assignmentId: parsedInput.assignmentId,
         userId,
         accessType: EnumAccessType.Write,
       });
@@ -32,38 +30,10 @@ export const addQuestion = createSafeActionClient()
         return { type: EnumAddQuestionResult.NotAuthorized };
       }
 
-      const newQuestionId = randomUUID();
-      const questionTextFileName = `questions/${newQuestionId}/question.txt`;
-      const questionTextBuffer = Buffer.from(question, "utf-8");
-      const didQuestionTextUploadSucceed = await putObject({
-        body: questionTextBuffer,
-        fileName: questionTextFileName,
-        contentType: "text/plain",
-      });
-
-      if (!didQuestionTextUploadSucceed) {
-        return { type: EnumAddQuestionResult.NotUploaded };
+      switch (parsedInput.type) {
+        case EnumQuestionType.Code:
+          return addCodeQuestion(parsedInput);
       }
-
-      const starterCodeFileName = `questions/${newQuestionId}/starterCode.txt`;
-      const starterCodeBuffer = Buffer.from(starterCode, "utf-8");
-      const didStarterCodeUploadSucceed = await putObject({
-        body: starterCodeBuffer,
-        fileName: starterCodeFileName,
-        contentType: "text/plain",
-      });
-
-      if (!didStarterCodeUploadSucceed) {
-        return { type: EnumAddQuestionResult.NotUploaded };
-      }
-
-      await addQuestionToDb({
-        questionId: newQuestionId,
-        name,
-        assignmentId,
-      });
-
-      return { type: EnumAddQuestionResult.QuestionAdded };
     } catch (e) {
       console.error(e);
       return { type: EnumAddQuestionResult.Error };
